:hammer: Refactor: Diary 디렉토리구조, 폴더명 변경
