+++ src/components/Diary/DiaryMainContainer/ChallengeSection/Challenge/AchievementRate/AchievementRate.scss
