+++ src/components/Diary/DiaryMainContainer/ChallengeSection/Challenge/Challenge.scss
