+++ src/components/Diary/DiaryMainContainer/ChallengeSection/Challenge/Goals/Goal.js
