+++ src/components/Diary/DiaryMainContainer/ChallengeSection/Challenge/Goals/GoalsList.scss
