+++ src/components/Diary/DiaryMainContainer/ChallengeSection/ChallengeSection.scss
