+++ src/components/Diary/DiaryMainContainer/SpecialNoteSection/SpecialNoteSection.js
