+++ src/components/Diary/DiaryMainContainer/WalkingSection/WalkingSection.js
