+++ src/components/Diary/DiaryMainContainer/WalkingSection/WalkingSection.scss
